allow extending NumpyDocString sections
Current implementation of NumpyDocString define sections in
`self._parsed_data` in `NumpyDocString.__init__()`, which is not
convenient for developers to extend this class to add more sections. By
setting section definitions as a class static variable, it will be more
convenient for developers to extend more sections by changing the class
static variable `sections`.