BUG: Make docscrape ParseError Python 3 compatible.
Fix `__str__` for `ParseError` in Python 3, which was resulting in error
messages with `numpydoc.docscrape.ParseError: <unprintable ParseError object>`
instead of the intended descriptive message.

Use `args[0]` instead of `Exception`'s `message` member, which had been removed
in Python 3. `args` is available to both Python 2 and 3. The `__str__` method
had been failing on an `AttributeError` when using `self.message`, resulting in
an unprintable Exception, which occluded the offending typo in a function name
which triggered the`ParseError`.