Clarify location of certain invalid docstrings.
If one adds the following invalid docstring at the end of
`doc/example.py`
```
def bar():
    """
    Parameters
    ----------
    a : int

    Parameters
    ----------
    b : float
    """
```
then as of numpydoc master the error message is
```
ValueError: The section Parameters appears twice in the docstring of None in None.
```

This patch fixes the error message to give
```
ValueError: The section Parameters appears twice in the docstring of <function bar at 0x7fa5c9d20e18> in /path/to/numpydoc/doc/example.py.
```