MAINT: rm redundant listing of section titles.
The validation module maintained an independent listing of valid
section modules. This information is already included in the
class attribute of NumpyDocString. This change improves maintainability
by removing this redundancy so that, if there is ever a change to the
standard in the future, it does not need to be repeated in multiple
locations in the source code.