ENH: Enable validation during sphinx-build process (#302)
* WIP: Move get_doc_obj to docscrape

* WIP: mv _obj property to NumpyDocString

* Proof-of-concept: Docstring attrs covered by refactor.

Running the test suite on this patch demonstrates that refactoring
the boundary between NumpyDocString and SphinxDocString provides the
necessary info to (potentially) do away with the validate.Docstring
class.

* NOTE TO SELF: get_doc_object in docscrape_sphinx

* Docstring -> Validator.

* Activate validation during sphinx-build.

Add a conf option to turn on/off.

TODO: test

* Replace logger.warn with warning.

logger.warn is apparently deprecated

* DOC: Add numpydoc_validate to conf docs.

* Add mechanism for validation check selection.

Adds a config option with a set to allow users to select
which validation checks are used. Default is an empty set,
which means none of the validation checks raise warnings
during the build process.

Add documentation for new option and activate in the doc build.

* TST: modify how MockApp sets builder app.

* TST: Add test of validation warnings.

* Specify some sensible validation defaults.

* Add docstring name to validation warnings.

* Add all keyword to validation_check configuration.

More flexibility in configuring which validation checks to run during
sphinx build. If 'all' is present, treat the rest of the set as a
blocklist, else an allowlist.

* Fix failing test.

* Make validation error mapping easier to read.

* Add check for invalid error codes in configuration.

plus test.

* Add feature to exclude patterns from docstring validation.

Modify updated config name to avoid sphinx warning.

Add documentation for exclusion config value.

* Be explicit about regex syntax for exclude config val

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>

* Rm redundant numpydoc_validate config param.

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>