BUG: Correct functionality of numpydoc SS05 (#613)
* BUG: Correct functionality of numpydoc SS05.

Currently SS05 checks for infinitive verbs by checking if the last character is "s". However, if the first word is "process", which is valid, this incorrectly throws a numpydoc validation error.

This commit allows the first word to end in two "s"es without flagging as an error. This will allow for words like "process" or "progress," while still checking for invalid words like "creates."

* Update numpydoc/validate.py

* FIX: Overrides

---------

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>