Alter _find_type_hint for better type inference
The _find_type_hint method now provides better type inference for class attributes without type hints at the cost of less useful types for the Methods section. However, that section does not seem to be using the types in the final Sphinx render and it is not clear that types even make sense there, so it might be a reasonable trade-off.