DEP: Break numpydoc.cli dependence on sphinx.
The only thing here that depends on sphinx is the
fn which uses docscrape_sphinx's version of get_doc_object.
If we switch to docscrape's version of get_doc_object, then the
resulting rendering is derived from NumpyDocString instead of
SphinxDocString.

Alternatively - this could be left as-is and the test_render tests
in  could be importorskipped.